pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.42",
    "werkzeug>=3.1.3",
    "sendgrid>=6.12.4",
    "telegram>=0.0.1",
    "openai>=1.99.6",
    "requests>=2.32.4",
    "python-telegram-bot[http2,socks]==20.3",
    "pandas>=2.3.1",
    "openpyxl>=3.1.5",
    "numpy>=2.3.2",
]
